% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R, R/learningfunctions.R
\name{AdjustFilter}
\alias{AdjustFilter}
\title{Adjust the filter such that it reflects the pooling operation performed.}
\usage{
AdjustFilter(poolingFilter, A.hat, X)

AdjustFilter(poolingFilter, A.hat, X)
}
\arguments{
\item{poolingFilter}{The original pooling filter for the model, contained
within the modelResults object.}

\item{A.hat}{The normalized Laplacian for convolution.}

\item{X}{The input predictions.}

\item{Theta}{Weight matrix from last training iteration.}
}
\description{
Adjust the filter such that it reflects the pooling operation performed.

Adjust the pooling filter (necessary for median, min, or max pooling).
}
