% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coregulationgraphfunctions.R
\name{BuildCoRegulationGraph}
\alias{BuildCoRegulationGraph}
\title{Construct a co-regulation graph where nodes are analytes, and edges
are present between nodes and edges that meet the processing cutoffs. 
The graph generated by this function is undirected and
is an igraph object.}
\usage{
BuildCoRegulationGraph(inputResults, vertexSize = 15)
}
\arguments{
\item{inputResults}{A data frame containing the analyte names for each model
and the coefficients, p-values, and R^2 values.}

\item{vertexSize}{Width of each vertex in pixels}
}
\description{
Construct a co-regulation graph where nodes are analytes, and edges
are present between nodes and edges that meet the processing cutoffs. 
The graph generated by this function is undirected and
is an igraph object.
}
