% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{CalculatePredictionCutoffs}
\alias{CalculatePredictionCutoffs}
\title{Calculate minimum and maximum cutoffs for prediction.}
\usage{
CalculatePredictionCutoffs(modelResults)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}
}
\description{
Calculate minimum and maximum cutoffs for prediction.
}
