% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{CombineSubspaces}
\alias{CombineSubspaces}
\title{Combine subspaces (i.e. using Grassmann Manifold Technique
- see PMID 30329022) given two different modalities of data (e.g. gene and metabolite)
and the alpha value and the desired number of eigenvectors.}
\usage{
CombineSubspaces(type1Similarity, type2Similarity, eigenCount, alpha)
}
\arguments{
\item{type1Similarity}{A cosine similarity matrix for the first data type, 
found using ComputeCosineSimilarity.}

\item{type2Similarity}{A cosine similarity matrix for the second data type,
found using ComputeCosineSimilarity.}

\item{eigenCount}{The number of eigenvectors to use.}

\item{alpha}{The value of alpha to use.}
}
\value{
A named list including the data projected onto the merged subspace,
the optimal number of eigenvectors, the optimal alpha value, the clustering
coefficient, and the dendrogram.
}
\description{
Combine subspaces (i.e. using Grassmann Manifold Technique
- see PMID 30329022) given two different modalities of data (e.g. gene and metabolite)
and the alpha value and the desired number of eigenvectors.
}
