% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{CompositePrediction}
\alias{CompositePrediction}
\title{Obtain a prediction from a composite model, given the pairs to include in the model.}
\usage{
CompositePrediction(
  pairs,
  modelResults,
  minCutoff,
  maxCutoff,
  useCutoff = FALSE
)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{modelResults}{A ModelResults object.}

\item{minCutoff}{Mininum cutoff for the prediction.}

\item{maxCutoff}{Maximum cutoff for the prediction.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}
}
\value{
A final predicted value for each sample in the input data
}
\description{
Obtain a prediction from a composite model, given the pairs to include in the model.
}
