% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{CompositePredictionLOO}
\alias{CompositePredictionLOO}
\title{Obtain a prediction from composite models with each of the individual pairs
removed, given}
\usage{
CompositePredictionLOO(pairs, IntLIMresults, modelResults, inputData)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{IntLIMresults}{The filtered results obtained from the IntLIM function
ProcessRe}

\item{modelResults}{A ModelResults object.}

\item{inputData}{An object of type IntLimData.}
}
\value{
A list of predicted values for each sample in the input data
}
\description{
Obtain a prediction from composite models with each of the individual pairs
removed, given
}
