% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{ComputeCosineSimilarity}
\alias{ComputeCosineSimilarity}
\title{Compute cosine similarity between samples on an adjacency matrix.}
\usage{
ComputeCosineSimilarity(R)
}
\arguments{
\item{R}{The adjacency matrix.}
}
\value{
A matrix of sample rows and sample columns, with the cosine
similarities listed as the values.
}
\description{
Compute cosine similarity between samples on an adjacency matrix.
}
