% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{ComputeGlobalErrorImportance}
\alias{ComputeGlobalErrorImportance}
\title{Computes the importance as the median absolute error for all predictors
on all samples except the sample in question.}
\usage{
ComputeGlobalErrorImportance(predictions, true)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{true}{Named vector of the true outcome values.}
}
\value{
A list of vectors (one for each sample) with an importance metric.
}
\description{
Computes the importance as the median absolute error for all predictors
on all samples except the sample in question.
}
