% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{ComputeGlobalErrorImportanceTest}
\alias{ComputeGlobalErrorImportanceTest}
\title{Computes the importance as the median absolute error for all predictors
samples in the training data.}
\usage{
ComputeGlobalErrorImportanceTest(predictions, true)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{true}{Named vector of the true outcome values.}
}
\value{
A data frame with the importance metric for each sample and each
predictor. For this metric, the value will be the same for all samples.
}
\description{
Computes the importance as the median absolute error for all predictors
samples in the training data.
}
