% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{ComputeGlobalPredImportance}
\alias{ComputeGlobalPredImportance}
\title{Computes the importance as the median absolute error for all predictors
on all samples except the sample in question.}
\usage{
ComputeGlobalPredImportance(predictions, true, outliers)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{true}{Named vector of the true outcome values.}

\item{outliers}{A vector of outliers for each sample. Generated using 
FindDistributionalOutliers().}
}
\value{
A list of vectors (one for each sample) with an importance metric.
}
\description{
Computes the importance as the median absolute error for all predictors
on all samples except the sample in question.
}
