% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{computeGradientImportanceOnly}
\alias{computeGradientImportanceOnly}
\title{Compute the gradient for a single layer neural network.}
\usage{
computeGradientImportanceOnly(modelResults, importance, Y.pred)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{importance}{Importance metrics for each predictor and sample.}

\item{Y.pred}{The predicted phenotype value.}
}
\description{
Compute the gradient for a single layer neural network.
}
