% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ComputeIndividualPerformance}
\alias{ComputeIndividualPerformance}
\title{Obtain a prediction from a composite model, given the pairs to include in the model.}
\usage{
ComputeIndividualPerformance(
  modelResults,
  trueVal,
  pruningMethod,
  useActivation = FALSE
)
}
\arguments{
\item{modelResults}{A ModelResults object.}

\item{trueVal}{The true values (predictions or outcomes) of the input data.}

\item{useActivation}{Whether or not to apply an activation function.}
}
\value{
A vector of significance values named by model.
}
\description{
Obtain a prediction from a composite model, given the pairs to include in the model.
}
