% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ComputeInfoGain}
\alias{ComputeInfoGain}
\title{Compute the information gain for a given prediction. The information gain
is computed as the difference between the entropy of the true values and the
conditional entropy of the true value given the prediction. Here, entropy
is defined using the probability that a value is greater than the midrange
of true values for the input data.}
\usage{
ComputeInfoGain(pred, trueVal)
}
\arguments{
\item{trueVal}{The true values (predictions or outcomes) of the input data.}

\item{pairs}{A list of pairs to include in the composite model.}
}
\description{
Compute the information gain for a given prediction. The information gain
is computed as the difference between the entropy of the true values and the
conditional entropy of the true value given the prediction. Here, entropy
is defined using the probability that a value is greater than the midrange
of true values for the input data.
}
