% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{ComputeLocalErrorImportanceTest}
\alias{ComputeLocalErrorImportanceTest}
\title{Computes the importance as the median absolute error for local predictors
(i.e. the predictions for k nearest neighbors of each sample).}
\usage{
ComputeLocalErrorImportanceTest(
  predictions,
  true,
  k,
  inputDataTrain,
  inputDataTest,
  eigStep = 10,
  alphaMin = 0,
  alphaMax = 1,
  alphaStep = 0.1
)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{true}{Named vector of the true outcome values.}

\item{k}{The number of nearest neighbors to consider.}

\item{inputDataTrain}{The training data (in IntLimData format).}

\item{inputDataTest}{The testing data (in IntLimData format).}

\item{eigStep}{The number of eigenvectors to step by during the evaluation.
Note that this must be less than the number of samples. Default = 10.}

\item{alphaMin}{The lowest value of alpha to investigate. Default = 0.}

\item{alphaMax}{The highest value of alpha to investigate. Default = 1.}

\item{alphaStep}{The value of alpha to step by during the evalutation.
Default = 0.1.}
}
\value{
A list of vectors (one for each sample) with an importance metric.
}
\description{
Computes the importance as the median absolute error for local predictors
(i.e. the predictions for k nearest neighbors of each sample).
}
