% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ComputeMetaFeatureWeights}
\alias{ComputeMetaFeatureWeights}
\title{Compute the weight of each predictor given the weights of different
metafeatures.}
\usage{
ComputeMetaFeatureWeights(metaFeatures, modelResults)
}
\arguments{
\item{metaFeatures}{A set of metafeatures, such as that found within ModelResults}

\item{currentWeights}{The weights for each metafeature}
}
\value{
A weight matrix for each sample and each predictor.
}
\description{
Compute the weight of each predictor given the weights of different
metafeatures.
}
