% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ComputeOddsRatio}
\alias{ComputeOddsRatio}
\title{Compute the odds for a given prediction. The odds ratio
is computed as the odds of predicting correctly (within a margin) using
the subgraph against correct prediction using the mean true prediction.}
\usage{
ComputeOddsRatio(pred, trueVal, margin = 0.1)
}
\arguments{
\item{trueVal}{The true values (predictions or outcomes) of the input data.}

\item{margin}{The margin of error for a prediction to be considered "correct".
Default is 0.1}

\item{pairs}{A list of pairs to include in the composite model.}
}
\description{
Compute the odds for a given prediction. The odds ratio
is computed as the odds of predicting correctly (within a margin) using
the subgraph against correct prediction using the mean true prediction.
}
