% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{ComputePathwayImportance}
\alias{ComputePathwayImportance}
\title{Compute the importance using pathway membership. Predictor pairs that share
at least one pathway have an importance of 1, and all others have an 
importance of 0.}
\usage{
ComputePathwayImportance(predictions, inputData, colIdInd, colIdOut)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{inputData}{Input data, which may include database ID mappings for
the analytes.}

\item{colIdInd}{The ID of the column that has the analyte ID's for the
independent variable. If blank, then the existing ID's are used.}

\item{colIdOut}{The ID of the column that has the analyte ID's for the
outcome variable. If blank, then the existing ID's are used.}
}
\value{
A data frame with the importance metric for each sample and each
predictor. The metric will be the same for all samples in this case.
}
\description{
Compute the importance using pathway membership. Predictor pairs that share
at least one pathway have an importance of 1, and all others have an 
importance of 0.
}
