% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{ComputePDFImportance}
\alias{ComputePDFImportance}
\title{Computes the importance as the density in the probability density function,
normalized over all densities.}
\usage{
ComputePDFImportance(predictions)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}
}
\value{
A list of vectors (one for each sample) with an importance metric.
}
\description{
Computes the importance as the density in the probability density function,
normalized over all densities.
}
