% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{ComputePvalImportance}
\alias{ComputePvalImportance}
\title{Compute the importance using p-value of model interaction term. Specifically,
the p-value percentile is calculated, and the weight is taken as 1 - the percentile,
so that the lowest p-values are weighted more heavily.}
\usage{
ComputePvalImportance(predictions, modelStats)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{modelStats}{A data frame that includes the interaction p-values and
interaction coefficients for each pair (such as the one output by IntLIM's
ProcessResults function)}
}
\value{
A list of vectors (one for each sample) with an importance metric.
}
\description{
Compute the importance using p-value of model interaction term. Specifically,
the p-value percentile is calculated, and the weight is taken as 1 - the percentile,
so that the lowest p-values are weighted more heavily.
}
