% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{ComputeReactionImportance}
\alias{ComputeReactionImportance}
\title{Compute the importance using reaction membership. Predictor pairs that share
at least one reaction have an importance of 1, and all others have an 
importance of 0. Outliers are also set to 0.}
\usage{
ComputeReactionImportance(
  predictions,
  analytehasreaction,
  analyteNamesOut,
  analyteNamesInd,
  source,
  outliers
)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{analytehasreaction}{A mapping from analytes to reactions. This
mapping has two columns: RaMPID and PathID}

\item{analyteNamesOut}{A data frame mapping the analyte identifier for the
outcome analyte from the IntLIM model to an identifier that can be mapped
to RaMP.}

\item{analyteNamesInd}{A data frame mapping the analyte identifier for the
independent variable analyte from the IntLIM model to an identifier that can be mapped
to RaMP.}

\item{source}{A data frame that maps analyte identifiers to RaMP internal
identifiers.}

\item{outliers}{A vector of outliers for each sample. Generated using 
FindDistributionalOutliers().}
}
\value{
A list of vectors (one for each sample) with an importance metric.
}
\description{
Compute the importance using reaction membership. Predictor pairs that share
at least one reaction have an importance of 1, and all others have an 
importance of 0. Outliers are also set to 0.
}
