% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ComputeTScore}
\alias{ComputeTScore}
\title{Compute t-score for a prediction using the paired t-test. The t-score
tells us how far away the prediction errors are from the mean errors.}
\usage{
ComputeTScore(pred, trueVal)
}
\arguments{
\item{trueVal}{The true values (predictions or outcomes) of the input data.}

\item{pairs}{A list of pairs to include in the composite model.}
}
\description{
Compute t-score for a prediction using the paired t-test. The t-score
tells us how far away the prediction errors are from the mean errors.
}
