% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringfunctions.R
\name{computeVariance}
\alias{computeVariance}
\title{Compute the variance within a cluster.}
\usage{
computeVariance(cluster, clusterName, modelInputs, allVariances)
}
\arguments{
\item{cluster}{A set of nodes.}

\item{clusterName}{The name assigned to a set of nodes.}

\item{modelInputs}{An object of the modelInputs class.}

\item{allVariances}{A list of all variances investigated, where each variance 
refers to a single cluster.}
}
\description{
Compute the variance within a cluster.
}
