% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{CosineSimilarityToTrainingOnSingleDataType}
\alias{CosineSimilarityToTrainingOnSingleDataType}
\title{Compute cosine similarity between a testing sample and each training sample
for a single data type.}
\usage{
CosineSimilarityToTrainingOnSingleDataType(trainingData, testingData, samp)
}
\arguments{
\item{trainingData}{The training data.}

\item{testingData}{The testing data.}

\item{samp}{The name of the testing sample against which to compute similarity.}

\item{opt}{The optimal model obtained using FindOptimalSubspaceClustering on
the training data.}
}
\value{
A matrix of similarities, where the rows are the training data samples
and the columns are the testing data samples.
}
\description{
Compute cosine similarity between a testing sample and each training sample
for a single data type.
}
