% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{CosineSimilarityToTrainingOnSubspace}
\alias{CosineSimilarityToTrainingOnSubspace}
\title{Compute cosine similarity between a testing sample and each training sample
in the combined subspace matrix.}
\usage{
CosineSimilarityToTrainingOnSubspace(opt, inputDataTest, inputDataTrain, samp)
}
\arguments{
\item{opt}{The optimal model obtained using FindOptimalSubspaceClustering on
the training data.}

\item{inputDataTest}{The testing data.}

\item{inputDataTrain}{The training data.}

\item{samp}{The name of the testing sample against which to compute similarity.}
}
\value{
A matrix of similarities, where the rows are the training data samples
and the columns are the testing data samples.
}
\description{
Compute cosine similarity between a testing sample and each training sample
in the combined subspace matrix.
}
