% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_AssignToPools.R
\name{CreateHierarchicalPoolingFilter}
\alias{CreateHierarchicalPoolingFilter}
\title{Create the graph pooling filter, given the adjacency matrix of the input graph.}
\usage{
CreateHierarchicalPoolingFilter(modelInputs, poolCount)
}
\arguments{
\item{modelInputs}{An object of type "ModelInput".}

\item{poolCount}{Number of pools.}
}
\description{
Create the graph pooling filter, given the adjacency matrix of the input graph.
}
