% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{CreatePoolingFilterAllFolds}
\alias{CreatePoolingFilterAllFolds}
\title{Wrapper for CreatePoolingFilter.}
\usage{
CreatePoolingFilterAllFolds(modelInputs, poolType)
}
\arguments{
\item{modelInputs}{A list of objects of type "ModelInputs".}

\item{poolType}{One of "mean", "median", "max", or "min".}
}
\description{
Wrapper for CreatePoolingFilter.
}
