% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{Dhat}
\alias{Dhat}
\title{Compute the D-hat value, which is the denominator value of the composite prediction
for a sample k and a subgraph lambda.}
\usage{
Dhat(Aind, phi, M, Beta2, Beta3)
}
\arguments{
\item{Aind}{The matrix containing the values of the independent analyte type.}

\item{phi}{The weights of the importance metrics.}

\item{M}{The importance values.}

\item{Beta2}{The phenotype term coefficients.}

\item{Beta3}{The interaction term coefficients.}
}
\description{
Compute the D-hat value, which is the denominator value of the composite prediction
for a sample k and a subgraph lambda.
}
