% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{DhatPrime}
\alias{DhatPrime}
\title{Compute the D-hat value, which is the derivative of the denominator value of the composite prediction
for a sample k, a subgraph lambda, and a meta-feature metric gamma.}
\usage{
DhatPrime(Aind, M, Beta2, Beta3)
}
\arguments{
\item{Aind}{The matrix containing the values of the independent analyte type.}

\item{M}{The meta-features.}

\item{Beta2}{The phenotype term coefficients.}

\item{Beta3}{The interaction term coefficients.}
}
\description{
Compute the D-hat value, which is the derivative of the denominator value of the composite prediction
for a sample k, a subgraph lambda, and a meta-feature metric gamma.
}
