% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_ModelSetup.R
\name{DoModelSetup}
\alias{DoModelSetup}
\title{This function performs the following tasks:
1. Run IntLIM to obtain all pairwise models.
2. Filter the IntLIM results to obtain a subset of models.
3. Predict the phenotype of all training data given each pairwise model.
4. Build the co-regulation graph using the set of pairwise models.
5. Initialize model with parameters.}
\usage{
DoModelSetup(
  inputData,
  stype,
  outcomeType = 1,
  independentVarType = 2,
  continuous = TRUE,
  pvalcutoff = 1,
  rsquaredCutoff = 0,
  interactionCoeffPercentile = 0,
  metaFeatureList = c("pdf", "interactionpval", "interactioncoef", "analytecoef",
    "localerr"),
  k = 2,
  eigStep = 1,
  colIdInd = "databaseId",
  colIdOut = "databaseId",
  edgeTypeList = c("shared.outcome.analyte", "shared.independent.analyte"),
  learningRate = 0.2,
  covar = c(),
  maxIterations = 1000,
  convergenceCutoff = 0.001,
  activationType = "none",
  optimizationType = "SGD",
  initialMetaFeatureWeights = 0
)
}
\arguments{
\item{inputData}{An object with the following fields:}

\item{stype}{The phenotype (outcome) to predict. This can be either a categorical
or numeric outcome.}

\item{outcomeType}{The outcome type (1 or 2)}

\item{independentVarType}{The independent variable type (1 or 2)}

\item{continuous}{Whether or not the outcome is continuous. Default is TRUE.}

\item{metaFeatureList}{A list of the valid metrics to include. Valid metrics are
"pdf", "localerr", "globalerr", and "pathway".}

\item{k}{The number of nearest neighbors to consider in localerr.}

\item{eigStep}{The number of eigenvectors to step by during the evaluation
in localerr.
Note that this must be less than the number of samples in localerr. Default = 10.}

\item{colIdInd}{The ID of the column that has the analyte ID's for the
independent variable. If blank, then the existing ID's are used.}

\item{colIdOut}{The ID of the column that has the analyte ID's for the
outcome variable. If blank, then the existing ID's are used.}

\item{edgeTypeList}{List containing one or more of the following to include
in the line graph:
- "shared.outcome.analyte"
- "shared.independent.analyte"
- "analyte.chain"}

\item{learningRate}{Learning rate to use during training. Default is 0.2}

\item{covar}{The clinical covariates to include in the model. These should be the same
covariates that were included when running the IntLIM linear models.}

\item{maxIterations}{Maximum number of iterations. Default is 1,000.}

\item{convergenceCutoff}{Cutoff for convergence. Default is 0.001.}

\item{activationType}{Activation function. May be "softmax", "sigmoid", 
"tanh", or "none". Default is "none", meaning stype.class is numeric.}

\item{optimizationType}{Type of optimization. May be "SGD", "momentum",
"adagrad", or "adam". Default is "SGD".}

\item{initialMetaFeatureWeights}{Initial weights for model meta-features. Default
is 0, which results in each meta-feature being given equal weight.}
}
\description{
This function performs the following tasks:
1. Run IntLIM to obtain all pairwise models.
2. Filter the IntLIM results to obtain a subset of models.
3. Predict the phenotype of all training data given each pairwise model.
4. Build the co-regulation graph using the set of pairwise models.
5. Initialize model with parameters.
}
