% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_AssignToPools.R
\name{DoPooling}
\alias{DoPooling}
\title{Create the graph pooling filter, given the adjacency matrix of the input graph.}
\usage{
DoPooling(modelInput, method = "kmeans", layers = 1, poolCount)
}
\arguments{
\item{modelInput}{An object of type "ModelInput".}

\item{method}{One of "kmeans", "hierarchical", "supernodes", or "betweenness".}

\item{poolCount}{Number of pools. If more than one layer of pooling is desired,
this parameter should be a vector where the first number is the number of
pools in the first layer.}
}
\description{
Create the graph pooling filter, given the adjacency matrix of the input graph.
}
