% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningfunctions.R
\name{DoPredictionImportanceOnly}
\alias{DoPredictionImportanceOnly}
\title{Predict Y given current weights.}
\usage{
DoPredictionImportanceOnly(modelResults, iteration, importance)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{iteration}{The current iteration.}

\item{importance}{Importance metrics for each predictor and sample.}
}
\description{
Predict Y given current weights.
}
