% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{DoSignificancePropagation}
\alias{DoSignificancePropagation}
\title{Obtain a prediction from a composite model, given the pairs to include in the model.}
\usage{
DoSignificancePropagation(
  pairs,
  modelResults,
  covar = NULL,
  verbose = FALSE,
  makePlots = FALSE
)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{modelResults}{A ModelResults object.}

\item{verbose}{Whether or not to print out each step.}

\item{makePlots}{Whether or not to plot the pruned model at each step.}
}
\value{
A final predicted value for each sample in the input data
}
\description{
Obtain a prediction from a composite model, given the pairs to include in the model.
}
