% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{DoSingleTrainingIteration}
\alias{DoSingleTrainingIteration}
\title{Train the graph learning model, using the specifications in the ModelResults
class and storing the results in the ModelResults class.}
\usage{
DoSingleTrainingIteration(
  modelResults,
  prunedModels,
  minCutoff,
  maxCutoff,
  useCutoff = FALSE
)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{prunedModels}{The models that remain after pruning.}

\item{minCutoff}{Mininum cutoff for the prediction.}

\item{maxCutoff}{Maximum cutoff for the prediction.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}
}
\description{
Train the graph learning model, using the specifications in the ModelResults
class and storing the results in the ModelResults class.
}
