% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningfunctions.R
\name{DoSingleTrainingIterationImportanceOnly}
\alias{DoSingleTrainingIterationImportanceOnly}
\title{Train the graph learning model, using the specifications in the ModelResults
class and storing the results in the ModelResults class.T}
\usage{
DoSingleTrainingIterationImportanceOnly(modelResults, iteration, importance)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{iteration}{The current iteration.}

\item{importance}{Importance metrics for each predictor and sample.}
}
\description{
Train the graph learning model, using the specifications in the ModelResults
class and storing the results in the ModelResults class.T
}
