% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_PredictUsingLearnedModel.R
\name{DoTestSetupAndPrediction}
\alias{DoTestSetupAndPrediction}
\title{Sets up the test data so that it is in the correct format, then
runs prediction on the test data. To do this:
1. Run pairwise prediction on the test data.
2. Run pairwise prediction on the training data.
3. Compute metafeatures on the test data.
4. Predict the test values using the learned model.}
\usage{
DoTestSetupAndPrediction(
  inputDataTest,
  model,
  k = 2,
  eigStep = 1,
  colIdInd = "databaseId",
  colIdOut = "databaseId",
  useCutoff = FALSE
)
}
\arguments{
\item{inputDataTest}{An object of the IntLimData class corresponding to the test set.}

\item{model}{An object of the ModelResults class corresponding to the optimized model.}

\item{k}{The number of nearest neighbors to consider in localerr.}

\item{eigStep}{The number of eigenvectors to step by during the evaluation
in localerr.
Note that this must be less than the number of samples in localerr. Default = 10.}

\item{colIdInd}{The ID of the column that has the analyte ID's for the
independent variable. If blank, then the existing ID's are used.}

\item{colIdOut}{The ID of the column that has the analyte ID's for the
outcome variable. If blank, then the existing ID's are used.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}
}
\value{
A vector of final prediction values for the test data.
}
\description{
Sets up the test data so that it is in the correct format, then
runs prediction on the test data. To do this:
1. Run pairwise prediction on the test data.
2. Run pairwise prediction on the training data.
3. Compute metafeatures on the test data.
4. Predict the test values using the learned model.
}
