% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{FindDistributionalOutliers}
\alias{FindDistributionalOutliers}
\title{Finds and returns the distributional outliers for each sample.}
\usage{
FindDistributionalOutliers(
  predictions,
  lowerPercentileLimit = 0.1,
  upperPercentileLimit = 0.9
)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{lowerPercentileLimit}{Sample-specific percentile below which outliers
will be removed.}

\item{upperPercentileLimit}{Sample-specific percentile above which outliers
will be removed.}
}
\value{
A list of vectors (one for each sample) where the outliers are masked
with a 1. Non-outliers are labeled as 0.
}
\description{
Finds and returns the distributional outliers for each sample.
}
