% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{FindOptimalSubspaceClustering}
\alias{FindOptimalSubspaceClustering}
\title{Finds the optimal subspace clustering (i.e. using Grassmann Manifold Technique
- see PMID 30329022) given two different modalities of data (e.g. gene and metabolite).
It optimizes the cophenetic correlation of the hierarchicial clustering
of the samples using a grid search.}
\usage{
FindOptimalSubspaceClustering(
  type1Similarity,
  type2Similarity,
  eigStep = 10,
  alphaMin = 0,
  alphaMax = 1,
  alphaStep = 0.1
)
}
\arguments{
\item{type1Similarity}{A cosine similarity matrix for the first data type, 
found using ComputeCosineSimilarity.}

\item{type2Similarity}{A cosine similarity matrix for the second data type,
found using ComputeCosineSimilarity.}

\item{eigStep}{The number of eigenvectors to step by during the evaluation.
Note that this must be less than the number of samples. Default = 10.}

\item{alphaMin}{The lowest value of alpha to investigate. Default = 0.}

\item{alphaMax}{The highest value of alpha to investigate. Default = 1.}

\item{alphaStep}{The value of alpha to step by during the evalutation.
Default = 0.1.}
}
\value{
A named list including the data projected onto the merged subspace,
the optimal number of eigenvectors, the optimal alpha value, the clustering
coefficient, and the dendrogram.
}
\description{
Finds the optimal subspace clustering (i.e. using Grassmann Manifold Technique
- see PMID 30329022) given two different modalities of data (e.g. gene and metabolite).
It optimizes the cophenetic correlation of the hierarchicial clustering
of the samples using a grid search.
}
