% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{FormatInput}
\alias{FormatInput}
\title{Format the input for graph-based learning. This input consists of:
1. The Laplacian of a line graph built from the co-regulation graphs, where 
each node corresponds to a pair of analytes.
2. A prediction value for each node of the line graph, for each sample X.
3. The true prediction values Y for each sample X.}
\usage{
FormatInput(
  predictionGraphs,
  coregulationGraph,
  importance,
  modelProperties,
  inputData,
  stype.class,
  edgeTypeList,
  stype,
  verbose = FALSE,
  covariates = list(),
  predictorBounds = c(NA, NA)
)
}
\arguments{
\item{predictionGraphs}{A list of igraph objects, each of which includes
predictions for each edge.}

\item{coregulationGraph}{An igraph object containing the coregulation graph.}

\item{stype.class}{The class of the outcome ("numeric" or "categorical")}

\item{edgeTypeList}{List containing one or more of the following to include
in the line graph:
- "shared.outcome.analyte"
- "shared.independent.analyte"
- "analyte.chain"}

\item{stype}{The phenotype or outcome of interest}

\item{verbose}{Whether to print the number of predictions replaced in each sample.
TRUE or FALSE. Default is FALSE.}

\item{covariates}{A list of covariates to include in the model. These will be in the
sampleMetaData slot of the inputData variable.}

\item{predictorBounds}{Upper and lower bounds for predictors. Predictors outside
of these bounds for a sample will not be included in the final prediction for that sample. 
Defaults to c(NA,NA), meaning that no predictors will be removed.}
}
\description{
Format the input for graph-based learning. This input consists of:
1. The Laplacian of a line graph built from the co-regulation graphs, where 
each node corresponds to a pair of analytes.
2. A prediction value for each node of the line graph, for each sample X.
3. The true prediction values Y for each sample X.
}
\section{Slots}{

\describe{
\item{\code{importance}}{A list of calculated importance metric data frames for each sample}

\item{\code{modelProperties}}{A data frame that includes model information, i.e. R^2,
interaction term p-value, and coefficients.}

\item{\code{inputData}}{An IntLIMData object that includes slots for the sample data,
the analyte data, and the analyte meta data.}
}}

