% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{GetAllImportanceMetrics}
\alias{GetAllImportanceMetrics}
\title{Finds and returns the distributional outliers for each sample.}
\usage{
GetAllImportanceMetrics(
  predictions,
  inputData,
  lowerPercentileLimit = 0.1,
  upperPercentileLimit = 0.9,
  metricList = c("pdf", "localerr", "globalerr", "pathway", "reaction",
    "interactionpval", "interactioncoef", "analytecoef"),
  modelStats = "",
  k = k,
  eigStep = 10,
  alphaMin = 0,
  alphaMax = 1,
  alphaStep = 0.1,
  analytehaspathway = "",
  reaction = "",
  analyteNamesOut = "",
  analyteNamesInd = "",
  source = "",
  stype = ""
)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{inputData}{The input data read in using the function IntLIM function
ReadData.}

\item{lowerPercentileLimit}{Sample-specific percentile below which outliers
will be removed.}

\item{upperPercentileLimit}{Sample-specific percentile above which outliers
will be removed.}

\item{metricList}{A list of the valid metrics to include. Valid metrics are
"pdf", "localerr", "globalerr", and "pathway".}

\item{modelStats}{A data frame that includes the interaction p-values and
interaction coefficients for each pair (such as the one output by IntLIM's
ProcessResults function)}

\item{k}{The number of nearest neighbors to consider in localerr.}

\item{eigStep}{The number of eigenvectors to step by during the evaluation
in localerr.
Note that this must be less than the number of samples in localerr. Default = 10.}

\item{alphaMin}{The lowest value of alpha to investigate in localerr. Default = 0.}

\item{alphaMax}{The highest value of alpha to investigate in localerr. Default = 1.}

\item{alphaStep}{The value of alpha to step by during the evaluation in localerr.
Default = 0.1.}

\item{analytehaspathway}{A mapping from analytes to pathways. This
mapping has two columns: RaMPID and PathID. Used in pathway importance.}

\item{reaction}{A list of analyte pairs with two columns: srcId and tgtId.}

\item{analyteNamesOut}{A data frame mapping the analyte identifier for the
outcome analyte from the IntLIM model to an identifier that can be mapped
to RaMP. Used in pathway importance.}

\item{analyteNamesInd}{A data frame mapping the analyte identifier for the
independent variable analyte from the IntLIM model to an identifier that can be mapped
to RaMP. Used in pathway importance.}

\item{source}{A data frame that maps analyte identifiers to RaMP internal
identifiers.}

\item{stype}{Phenotype or outcome to use in models.}
}
\value{
A list of data frames (one for each sample) with predictor importance
measured according to the listed criteria (one column per metric, one row
per predictor).
}
\description{
Finds and returns the distributional outliers for each sample.
}
