% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteringfunctions.R
\name{getLowestVarianceParent}
\alias{getLowestVarianceParent}
\title{Find the lowest variance parent of any of the current clusters.}
\usage{
getLowestVarianceParent(
  clusters,
  modelInputs,
  merges,
  labels,
  allClusters,
  allVariances
)
}
\arguments{
\item{clusters}{A list of clusters, where each cluster is a set of nodes.}

\item{modelInputs}{An object of the modelInputs class.}

\item{merges}{The merge slot of hierarchical clustering results as performed using
the hclust() function.}

\item{labels}{The labels for the numeric nodes in the hierarchical clustering
algorithm.}

\item{allClusters}{A list of all clusters investigated, where each cluster is
a set of nodes.}

\item{allVariances}{A list of all variances investigated, where each variance 
refers to a single cluster.}
}
\description{
Find the lowest variance parent of any of the current clusters.
}
