% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafeaturefunctions.R
\name{GetPredictorIDs}
\alias{GetPredictorIDs}
\title{Function for obtaining database identifiers for each predictor. Used
in functional analysis.}
\usage{
GetPredictorIDs(preds, inputData, colIdInd, colIdOut)
}
\arguments{
\item{inputData}{Input data, which may include database ID mappings for
the analytes.}

\item{colIdInd}{The ID of the column that has the analyte ID's for the
independent variable. If blank, then the existing ID's are used.}

\item{colIdOut}{The ID of the column that has the analyte ID's for the
outcome variable. If blank, then the existing ID's are used.}

\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}
}
\value{
A data frame with two columns. The indId column contains the ID's
for the independent analyte, and the outId column contains the ID's for
the outcome analyte.
}
\description{
Function for obtaining database identifiers for each predictor. Used
in functional analysis.
}
