% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{GetTestMetaFeatures}
\alias{GetTestMetaFeatures}
\title{Computes the metafeatures for each sample and model in the testing data.}
\usage{
GetTestMetaFeatures(
  predictionsTrain,
  predictionsTest,
  inputDataTrain,
  inputDataTest,
  lowerPercentileLimit = 0.1,
  upperPercentileLimit = 0.9,
  metaFeatureList = c("pdf", "localerr", "globalerr", "pathway", "reaction",
    "interactionpval", "interactioncoef", "analytecoef"),
  modelStats = "",
  k = k,
  eigStep = 10,
  alphaMin = 0,
  alphaMax = 1,
  alphaStep = 0.1,
  analyteNamesOut = "",
  analyteNamesInd = "",
  stype = "",
  colIdInd = "",
  colIdOut = ""
)
}
\arguments{
\item{predictionsTrain}{Prediction data frame for training data, where rows are samples, and
columns are predictors.}

\item{predictionsTest}{Prediction data frame for testing data, where rows are samples, and
columns are predictors.}

\item{inputDataTrain}{The training data (in IntLimData format).}

\item{inputDataTest}{The testing data (in IntLimData format).}

\item{modelStats}{A data frame that includes the interaction p-values and
interaction coefficients for each pair (such as the one output by IntLIM's
ProcessResults function)}

\item{k}{The number of nearest neighbors to consider in localerr.}

\item{eigStep}{The number of eigenvectors to step by during the evaluation
in localerr.
Note that this must be less than the number of samples in localerr. Default = 10.}

\item{alphaMin}{The lowest value of alpha to investigate in localerr. Default = 0.}

\item{alphaMax}{The highest value of alpha to investigate in localerr. Default = 1.}

\item{alphaStep}{The value of alpha to step by during the evaluation in localerr.
Default = 0.1.}

\item{analyteNamesOut}{A data frame mapping the analyte identifier for the
outcome analyte from the IntLIM model to an identifier that can be mapped
to RaMP. Used in pathway importance.}

\item{analyteNamesInd}{A data frame mapping the analyte identifier for the
independent variable analyte from the IntLIM model to an identifier that can be mapped
to RaMP. Used in pathway importance.}

\item{stype}{Phenotype or outcome to use in models.}

\item{colIdInd}{The ID of the column that has the analyte ID's for the
independent variable. If blank, then the existing ID's are used.}

\item{colIdOut}{The ID of the column that has the analyte ID's for the
outcome variable. If blank, then the existing ID's are used.}

\item{metricList}{A list of the valid metrics to include. Valid metrics are
"pdf", "localerr", "globalerr", and "pathway".}
}
\value{
A list of data frames (one for each sample) with predictor importance
measured according to the listed criteria (one column per metric, one row
per predictor).
}
\description{
Computes the metafeatures for each sample and model in the testing data.
}
