% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{InitializeGraphLearningModelAllFolds}
\alias{InitializeGraphLearningModelAllFolds}
\title{Wrapper for InitializeGraphLearningModel}
\usage{
InitializeGraphLearningModelAllFolds(
  modelInputs,
  poolingFilter,
  iterations = 1000,
  convergenceCutoff = 0.001,
  stype.class = "numeric",
  learningRate = 0.2,
  activationType = "softmax",
  weightsAfterPooling = FALSE,
  gradientType = "stochastic",
  optimizationType = "adam"
)
}
\arguments{
\item{modelInputs}{An object of the ModelInput class.}

\item{poolingFilter}{A matrix that pools convolution results.}

\item{iterations}{Maximum number of iterations. Default is 1,000.}

\item{convergenceCutoff}{Cutoff for convergence. Default is 0.001.}

\item{stype.class}{One of either "character" or "numeric". Default is "numeric".}

\item{learningRate}{Learning rate to use during training. Default is 0.2}

\item{activationType}{Activation function. May be "softmax", "sigmoid", 
or "tanh". Default is "softmax". Irrelevant if stype.class is numeric.}

\item{weightsAfterPooling}{Whether to include the weights after the pooling.
Default is FALSE.}

\item{gradientType}{One of either "stochastic" or "batch". Default is
"stochastic".}

\item{optimizationType}{Type of optimization. May be "gradient", "momentum",
"adagrad", or "adam". Default is "adam".}
}
\description{
Wrapper for InitializeGraphLearningModel
}
