% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{ModelInput-class}
\alias{ModelInput-class}
\title{ModelInput class}
\description{
ModelInput class
}
\section{Slots}{

\describe{
\item{\code{edge.wise.prediction}}{graph of co-regulation results from IntLIM}

\item{\code{true.phenotypes}}{data frame of true phenotypes given analyte pairs
and clinical covariates}

\item{\code{coregulation.graph}}{the original co-regulation graph for the input data.}

\item{\code{metaFeatures}}{A list of calculated meta-feature data frames for each sample}

\item{\code{model.properties}}{A data frame that includes model information, i.e. R^2,
interaction term p-value, and coefficients.}

\item{\code{input.data}}{An IntLIMData object that includes slots for the sample data,
the analyte data, and the analyte meta data.}

\item{\code{covariates}}{A list of covariates to use in the models.}

\item{\code{stype}}{Outcome/phenotype column name.}

\item{\code{stype.class}}{Class of outcome/phenotype (either "numeric" or "character").}
}}

