% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{Nhat}
\alias{Nhat}
\title{Compute the N-hat value, which is the numerator value of the composite prediction
for a sample k and a subgraph lambda.}
\usage{
Nhat(Aind, Aout, phi, M, Beta0, Beta1, BetaC, C)
}
\arguments{
\item{Aind}{The matrix containing the values of the independent analyte type.}

\item{Aout}{The matrix containing the values of the outcome analyte type.}

\item{phi}{The weights of the importance metrics.}

\item{M}{The importance values.}

\item{Beta0}{The intercept coefficients.}

\item{Beta1}{The analyte coefficients.}

\item{BetaC}{The covariate coefficients.}

\item{C}{The covariates for the predictors of interest. If no covariates,
this value should be NA.}
}
\description{
Compute the N-hat value, which is the numerator value of the composite prediction
for a sample k and a subgraph lambda.
}
