% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{NhatPrime}
\alias{NhatPrime}
\title{Compute the N-hat value, which is the derivative of the numerator value of the composite prediction
for a sample k, a subgraph lambda, and an importance metric gamma.}
\usage{
NhatPrime(Aind, Aout, M, Beta0, Beta1, BetaC, C)
}
\arguments{
\item{Aind}{The matrix containing the values of the independent analyte type.}

\item{Aout}{The matrix containing the values of the outcome analyte type.}

\item{M}{The importance values.}

\item{Beta0}{The intercept coefficients.}

\item{Beta1}{The analyte coefficients.}

\item{BetaC}{The covariate coefficients.}

\item{C}{The covariates for the predictors of interest.}
}
\description{
Compute the N-hat value, which is the derivative of the numerator value of the composite prediction
for a sample k, a subgraph lambda, and an importance metric gamma.
}
