% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ObtainCompositeModelsHiddenLayers}
\alias{ObtainCompositeModelsHiddenLayers}
\title{Obtain the list of pairs in a composite model.}
\usage{
ObtainCompositeModelsHiddenLayers(pairsInEachPredictor, importantPairs)
}
\arguments{
\item{pairsInEachPredictor}{A list of pairs contained within each predictor.}

\item{importantPairs}{A list of all pairs that were found to be important in the
previous layer.}
}
\description{
Obtain the list of pairs in a composite model.
}
