% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ObtainCompositeSubgraphs}
\alias{ObtainCompositeSubgraphs}
\title{Obtain the list of pairs in a composite model.}
\usage{
ObtainCompositeSubgraphs(modelInput)
}
\arguments{
\item{modelInput}{A ModelInput object}
}
\description{
Obtain the list of pairs in a composite model.
}
