% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ObtainSubgraphNeighborhoods}
\alias{ObtainSubgraphNeighborhoods}
\title{Obtain the list of pairs in a composite model.}
\usage{
ObtainSubgraphNeighborhoods(modelInput)
}
\arguments{
\item{modelInput}{A ModelInput object}
}
\value{
A list of sets, where each set is a neighborhood of nodes.
}
\description{
Obtain the list of pairs in a composite model.
}
