% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisepredictionfunctions.R
\name{OneHotEncoding}
\alias{OneHotEncoding}
\title{Perform one-hot encoding of the sample metadata. This is useful when categorical
covariates are used. This one-hot encoding will be used later in predictions,
and will simplify the prediction process because the names of the result columns
and the covariates will match up.}
\usage{
OneHotEncoding(inputData, covar = c())
}
\arguments{
\item{inputData}{An object with the following fields:}

\item{covar}{The clinical covariates to include in the model. These should be the same
covariates that were included when running the IntLIM linear models.}
}
\value{
A list with two elements: (1) the new set of covariates, and (2)
the new sample meta data, as encoded.
}
\description{
Perform one-hot encoding of the sample metadata. This is useful when categorical
covariates are used. This one-hot encoding will be used later in predictions,
and will simplify the prediction process because the names of the result columns
and the covariates will match up.
}
