% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{OptimizeImportanceCombo}
\alias{OptimizeImportanceCombo}
\title{Optimize the combination of predictors by importance alone (in other words,
exclude pooling and combine in a single layer using a linear combination).}
\usage{
OptimizeImportanceCombo(modelResults, verbose = TRUE)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{verbose}{Whether to print results as you run the model.}
}
\description{
Optimize the combination of predictors by importance alone (in other words,
exclude pooling and combine in a single layer using a linear combination).
}
