% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotCoRegulationGraphAllFolds}
\alias{PlotCoRegulationGraphAllFolds}
\title{Wrapper for PlotCoRegulationGraph.}
\usage{
PlotCoRegulationGraphAllFolds(
  graph,
  saveInDir = "",
  vertices = c(),
  truncateTo = 4
)
}
\arguments{
\item{graph}{The co-regulation graph.}

\item{saveInDir}{Directory where file should be saved. If "", then the output
is plotted without being saved. Default is ""}

\item{vertices}{List of vectors of vertices to plot. This is used if one
wishes to focus on a subset of vertices. If c(), then all vertices are plotted.
Default is c()}

\item{truncateTo}{Vertex names are truncated to the first "truncateTo" characters.
Default is 4. If -1, names are not truncated.}
}
\description{
Wrapper for PlotCoRegulationGraph.
}
