% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotGraphPredictionsAllFolds}
\alias{PlotGraphPredictionsAllFolds}
\title{Wrapper for PlotGraphPredictions.}
\usage{
PlotGraphPredictionsAllFolds(
  graphs,
  inputDataFolds,
  saveInDir = "",
  vertices = c(),
  truncateTo = 4
)
}
\arguments{
\item{graphs}{The graph with predictions projected onto it.}

\item{inputDataFolds}{List of named lists (output of 
CreateCrossValidationFolds()) with gene expression, metabolite abundances, 
and associated meta-data}

\item{saveInDir}{Directory where file should be saved. If "", then the output
is plotted without being saved. Default is "".}

\item{vertices}{List of vectors of vertices to plot. This is used if one
wishes to focus on a subset of vertices. If c(), then all vertices are plotted.
Default is c().}

\item{truncateTo}{Vertex names are truncated to the first "truncateTo" characters.
Default is 4. If -1, names are not truncated.}
}
\description{
Wrapper for PlotGraphPredictions.
}
