% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotLineGraph}
\alias{PlotLineGraph}
\title{Plot the line graph for each sample with nodes colored according to prediction.
Include a color scale for the predictions.}
\usage{
PlotLineGraph(
  modelResults,
  stype,
  subset = NULL,
  saveInDir = NULL,
  truncateTo = 2,
  weights = FALSE,
  analytes = NULL,
  includeLabels = TRUE,
  cutoffs = NULL,
  vertexSize = 10,
  sampSubset = NULL
)
}
\arguments{
\item{modelResults}{A ModelResults object.}

\item{stype}{Outcome / phenotype}

\item{subset}{The subset of predictors to include. If NULL, then all nodes
will be plotted.}

\item{saveInDir}{Directory where file should be saved. If NULL, then the output
is plotted without being saved. Default is NULL.}

\item{truncateTo}{Analyte names are truncated to the first "truncateTo" characters.
Default is 4. If NULL, names are not truncated.}

\item{weights}{Whether or not to encode weights in the opacity of the nodes.
Defaults to FALSE.}

\item{analytes}{analytes of interest. Default is NULL}

\item{includeLabels}{whether or not to include labels. Defaults to TRUE.}

\item{cutoffs}{Cutoff weight values, which can be included for visibility.
Default is NULL.}

\item{vertexSize}{Vertex size to use in display.}

\item{sampSubset}{Samples to plot.}
}
\description{
Plot the line graph for each sample with nodes colored according to prediction.
Include a color scale for the predictions.
}
