% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotLineGraphAllFolds}
\alias{PlotLineGraphAllFolds}
\title{Wrapper for PlotLineGraph.}
\usage{
PlotLineGraphAllFolds(
  modelInputs,
  stype,
  saveInDir = "",
  analytes = c(),
  truncateTo = 2,
  weights = c()
)
}
\arguments{
\item{modelInputs}{A list of ModelInput objects.}

\item{stype}{Outcome / phenotype}

\item{saveInDir}{Directory where file should be saved. If "", then the output
is plotted without being saved. Default is "".}

\item{analytes}{List of vectors of analytes to plot. This is used if one
wishes to focus on a subset of analytes If c(), then all vertices are plotted.
Default is c().}

\item{truncateTo}{Analyte names are truncated to the first "truncateTo" characters.
Default is 2. If c(), names are not truncated.}

\item{weights}{The list of weights assigned to each node. This is encoded using opacity.
If c(), all nodes are opaque. Default is c().}
}
\description{
Wrapper for PlotLineGraph.
}
