% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotPredictionDendrogram}
\alias{PlotPredictionDendrogram}
\title{Plots a color-coded dendrogram of predictions for a given sample. The
dendrogram is defined by the line graph, and the colors correspond to the
prediction made by each pair.}
\usage{
PlotPredictionDendrogram(
  modelInput,
  hierarchicalClustering,
  sampleIndex,
  predictionLimits
)
}
\arguments{
\item{modelInput}{A ModelInput object.}

\item{hierarchicalClustering}{The output of the doHierarchicalClustering
function.}

\item{predictionLimits}{A vector of the upper and lower limits to plot.
Predictions falling outside of this range will be modified to fall within
this range. This is only for the purposes of plotting, so that outlier
predictions do not obscure differences within the rest of the plot.}
}
\description{
Plots a color-coded dendrogram of predictions for a given sample. The
dendrogram is defined by the line graph, and the colors correspond to the
prediction made by each pair.
}
