% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotSubspaceClusteringHeatmap}
\alias{PlotSubspaceClusteringHeatmap}
\title{Plots a heatmap of the optimal subspace clustering.}
\usage{
PlotSubspaceClusteringHeatmap(optimalClustering)
}
\arguments{
\item{optimalClustering}{The output of FindOptimalSubspaceClustering.}
}
\description{
Plots a heatmap of the optimal subspace clustering.
}
