% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotWeightHeatmap}
\alias{PlotWeightHeatmap}
\title{Plot the current weight associated with each predictor and sample.}
\usage{
PlotWeightHeatmap(modelResults)
}
\arguments{
\item{modelResults}{A ModelResults object.}
}
\description{
Plot the current weight associated with each predictor and sample.
}
