% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_PredictUsingLearnedModel.R
\name{Predict}
\alias{Predict}
\title{Run a prediction on new data using the graph learning model.}
\usage{
Predict(
  pairs,
  targets,
  sources,
  inputData,
  weights,
  model,
  minCutoff,
  maxCutoff,
  useCutoff = FALSE,
  useActivation = TRUE,
  independentVarType,
  outcomeType
)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{targets}{Target analytes for all pairs}

\item{sources}{Source analytes for all pairs}

\item{inputData}{The input testing data, which should be of an 
IntLimData class type.}

\item{weights}{The weights of each model, computed using ComputeMetaFeatureWeights()}

\item{model}{The learned model, an object of the modelResults class.}

\item{minCutoff}{Mininum cutoff for the prediction.}

\item{maxCutoff}{Maximum cutoff for the prediction.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}

\item{useActivation}{Whether or not to use the activation function if the phenotype
to predict is a factor. Default is TRUE. We set to FALSE during training.}

\item{independentVarType}{The independent variable type (1 or 2)}

\item{outcomeType}{The outcome type (1 or 2)}
}
\value{
A vector of predictions
}
\description{
Run a prediction on new data using the graph learning model.
}
