% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{PredictFromLeastSquares}
\alias{PredictFromLeastSquares}
\title{Predict test data using a least squares model. This function
assumes that the weights are applied after pooling.}
\usage{
PredictFromLeastSquares(modelInput, trainingModelResults, leastSquaresModel)
}
\arguments{
\item{modelInput}{An object of the ModelInput class. This input corresponds
to the test data.}

\item{trainingModelResults}{An object of the ModelResults class. This corresponds to
the modelResults object for the training data.}

\item{leastSquaresModel}{A least squares model learned using 
RunLeastSquaresOptimization.}
}
\description{
Predict test data using a least squares model. This function
assumes that the weights are applied after pooling.
}
