% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{PredictFromLeastSquaresAllFolds}
\alias{PredictFromLeastSquaresAllFolds}
\title{Wrapper for RunLeastSquaresOptimization.}
\usage{
PredictFromLeastSquaresAllFolds(
  modelInput,
  trainingModelResults,
  leastSquaresModel
)
}
\arguments{
\item{modelInput}{A list of objects of the ModelInput class. This input corresponds
to the test data.}

\item{trainingModelResults}{A list of objects of the ModelResults class. This corresponds to
the modelResults object for the training data.}

\item{leastSquaresModel}{A list of least squares models learned using 
RunLeastSquaresOptimization.}
}
\description{
Wrapper for RunLeastSquaresOptimization.
}
