% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{PrunePredictorsHiddenLayers}
\alias{PrunePredictorsHiddenLayers}
\title{Given multiple composite predictors, prune the predictors that are not needed.}
\usage{
PrunePredictorsHiddenLayers(pairs, IntLIMresults, modelResults, inputData)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{IntLIMresults}{The filtered results obtained from the IntLIM function}

\item{modelResults}{A ModelResults object.}

\item{inputData}{An object of type IntLimData.}
}
\value{
A list of informative pairs
}
\description{
Given multiple composite predictors, prune the predictors that are not needed.
}
