% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{RunLeastSquaresOptimization}
\alias{RunLeastSquaresOptimization}
\title{Run least squares optimization to optimize the weights. This function
assumes that the weights are applied after pooling.}
\usage{
RunLeastSquaresOptimization(modelResults)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}
}
\description{
Run least squares optimization to optimize the weights. This function
assumes that the weights are applied after pooling.
}
