% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{RunLeastSquaresOptimizationAllFolds}
\alias{RunLeastSquaresOptimizationAllFolds}
\title{Wrapper for RunLeastSquaresOptimization.}
\usage{
RunLeastSquaresOptimizationAllFolds(modelResults)
}
\arguments{
\item{modelResults}{A list of objects of the ModelResults class.}
}
\description{
Wrapper for RunLeastSquaresOptimization.
}
