% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisepredictionfunctions.R
\name{RunPairwisePrediction}
\alias{RunPairwisePrediction}
\title{Given each significant pairwise model and the input data, predict the phenotype
for each sample. Recall that IntLIM models take the following form:
m ~ beta0 + beta1(g) + beta2(phenotype) + beta3(g:phenotype) + beta4...n(covariates)
Therefore, to predict phenotype given the betas learned by IntLIM, we use the
following model:
p ~ (m - (beta0 + beta1(g) + beta4...n(covariates)) / (beta2 + beta3(g))}
\usage{
RunPairwisePrediction(
  inputResults,
  inputData,
  stype = "",
  covar = c(),
  independentVarType = 2,
  outcomeType = 1
)
}
\arguments{
\item{inputResults}{The data frame of filtered results from IntLIM
model and processing results (output of ProcessResults()). All results must
include learned covariate weights (i.e. must be run with save.covar.pvals = TRUE)}

\item{inputData}{An object with the following fields:}

\item{stype}{The phenotype (outcome) to predict. This can be either a categorical
or numeric outcome.}

\item{covar}{The clinical covariates to include in the model. These should be the same
covariates that were included when running the IntLIM linear models.}

\item{independentVarType}{The independent variable type (1 or 2)}

\item{outcomeType}{The outcome type (1 or 2)}
}
\description{
Given each significant pairwise model and the input data, predict the phenotype
for each sample. Recall that IntLIM models take the following form:
m ~ beta0 + beta1(g) + beta2(phenotype) + beta3(g:phenotype) + beta4...n(covariates)
Therefore, to predict phenotype given the betas learned by IntLIM, we use the
following model:
p ~ (m - (beta0 + beta1(g) + beta4...n(covariates)) / (beta2 + beta3(g))
}
