% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{Backpropagate}
\alias{Backpropagate}
\title{Run backpropagation for a single layer. In backpropagation, a gradient is
is computed by taking partial derivatives for each of the model weights. Given
a learning rate, the weights are adjusted according to the gradient.}
\usage{
Backpropagate(modelResults, prunedModels, Y.pred)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{prunedModels}{The models that remain after pruning.}

\item{Y.pred}{The predicted phenotype value.}
}
\description{
Run backpropagation for a single layer. In backpropagation, a gradient is
is computed by taking partial derivatives for each of the model weights. Given
a learning rate, the weights are adjusted according to the gradient.
}
