% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{CompositePrediction}
\alias{CompositePrediction}
\title{Obtain a prediction from a composite model, given the pairs to include in the model.}
\usage{
CompositePrediction(pairs, IntLIMresults, modelResults, inputData)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{IntLIMresults}{The filtered results obtained from the IntLIM function
ProcessRe}

\item{modelResults}{A ModelResults object.}

\item{inputData}{An object of type IntLimData.}
}
\value{
A final predicted value for each sample in the input data
}
\description{
Obtain a prediction from a composite model, given the pairs to include in the model.
}
