% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{computeGradient}
\alias{computeGradient}
\title{Compute the gradient for a single layer neural network.}
\usage{
computeGradient(modelResults, prunedModels)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{prunedModels}{The models that remain after pruning.}
}
\description{
Compute the gradient for a single layer neural network.
}
