% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ComputeImportanceWeights}
\alias{ComputeImportanceWeights}
\title{Compute the weight of each predictor given the weights of different
importance metrics.}
\usage{
ComputeImportanceWeights(modelResults)
}
\arguments{
\item{modelResults}{A ModelResults object.}
}
\value{
A list of informative pairs
}
\description{
Compute the weight of each predictor given the weights of different
importance metrics.
}
