% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{ComputeInteractionCoefImportance}
\alias{ComputeInteractionCoefImportance}
\title{Compute the importance using absolute value of interaction term coefficient. Specifically,
the coefficient percentile is calculated and taken as the weight.}
\usage{
ComputeInteractionCoefImportance(predictions, modelStats)
}
\arguments{
\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}

\item{modelStats}{A data frame that includes the interaction p-values and
interaction coefficients for each pair (such as the one output by IntLIM's
ProcessResults function)}
}
\value{
A list of vectors (one for each sample) with an importance metric.
}
\description{
Compute the importance using absolute value of interaction term coefficient. Specifically,
the coefficient percentile is calculated and taken as the weight.
}
