% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ComputeMetaFeatureWeights}
\alias{ComputeMetaFeatureWeights}
\title{Compute the weight of each predictor given the weights of different
metafeatures.}
\usage{
ComputeMetaFeatureWeights(modelResults)
}
\arguments{
\item{modelResults}{A ModelResults object.}
}
\value{
A weight matrix for each sample and each predictor.
}
\description{
Compute the weight of each predictor given the weights of different
metafeatures.
}
