% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientfunctions.R
\name{DerivativeOfActivation}
\alias{DerivativeOfActivation}
\title{Compute the derivative of the error by the predictor.}
\usage{
DerivativeOfActivation(activationType, pred)
}
\arguments{
\item{activationType}{The name of the activation function.}

\item{pred}{The predicted value of a single sample for a single parameter
of interest.}

\item{importance}{The importance values.}
}
\description{
Compute the derivative of the error by the predictor.
}
