% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainingfunctions.R
\name{DoPrediction}
\alias{DoPrediction}
\title{Predict Y given current weights.}
\usage{
DoPrediction(modelResults, prunedModels)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{prunedModels}{The models that remain after pruning. There should only
be one model at the end.}
}
\description{
Predict Y given current weights.
}
