% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{DoSignificancePropagation}
\alias{DoSignificancePropagation}
\title{Obtain a prediction from a composite model, given the pairs to include in the model.}
\usage{
DoSignificancePropagation(
  pairs,
  modelResults,
  covar = c(),
  verbose = FALSE,
  pruningMethod = "error.t.test",
  modelRetention = "stringent",
  minCutoff,
  maxCutoff,
  useCutoff = FALSE
)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{modelResults}{A ModelResults object.}

\item{verbose}{Whether or not to print out each step.}

\item{pruningMethod}{The method to use for pruning. Right now, only "error.t.test" is valid.}

\item{modelRetention}{Strategy for model retention. "stringent" (the default)
retains only models that improve the prediction score. "lenient" also retains models that
neither improve nor reduce the prediction score.}

\item{minCutoff}{Mininum cutoff for the prediction.}

\item{maxCutoff}{Maximum cutoff for the prediction.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}
}
\value{
A final predicted value for each sample in the input data
}
\description{
Obtain a prediction from a composite model, given the pairs to include in the model.
}
