% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_LearnGraphPredictionModel.R
\name{FormatInput}
\alias{FormatInput}
\title{Format the input for graph-based learning. This input consists of:
- The Laplacian of a line graph built from the co-regulation graphs, where 
each node corresponds to a pair of analytes.
- A prediction value for each node of the line graph, for each sample X.
- The true prediction values Y for each sample X.
- The co-regulation graph.
- The line graph built from the co-regulation graphs.
- The input data (analyte levels, covariates, and phenotype).
- The model results from IntLIM for each predictor.
- The metafeatures for each model from IntLIM.
- The phenotype / outcome column name in the input data.
- The type of the outcome / phenotype ("numeric" or "categorical").
- The analyte type used as the outcome variable for each model (1 or 2).
- The analyte type used as the independent variable for each model (1 or 2).}
\usage{
FormatInput(
  predictionGraphs,
  coregulationGraph,
  metaFeatures,
  modelProperties,
  inputData,
  stype.class,
  edgeTypeList,
  stype,
  verbose = FALSE,
  covariates = list(),
  outcome = 2,
  independent.var.type = 2
)
}
\arguments{
\item{predictionGraphs}{A list of igraph objects, each of which includes
predictions for each edge.}

\item{coregulationGraph}{An igraph object containing the coregulation graph.}

\item{stype.class}{The class of the outcome ("numeric" or "categorical")}

\item{edgeTypeList}{List containing one or more of the following to include
in the line graph:
- "shared.outcome.analyte"
- "shared.independent.analyte"
- "analyte.chain"}

\item{stype}{The phenotype or outcome of interest}

\item{verbose}{Whether to print the number of predictions replaced in each sample.
TRUE or FALSE. Default is FALSE.}

\item{covariates}{A list of covariates to include in the model. These will be in the
sampleMetaData slot of the inputData variable.}

\item{outcome}{The outcome used in the IntLIM models.}

\item{independent.var.type}{The independent variable type used in the IntLIM models.}
}
\value{
An object of the ModelInput class. This object will contain the input data,
graph information, and individual predictor properties and their metafeatures.
}
\description{
Format the input for graph-based learning. This input consists of:
- The Laplacian of a line graph built from the co-regulation graphs, where 
each node corresponds to a pair of analytes.
- A prediction value for each node of the line graph, for each sample X.
- The true prediction values Y for each sample X.
- The co-regulation graph.
- The line graph built from the co-regulation graphs.
- The input data (analyte levels, covariates, and phenotype).
- The model results from IntLIM for each predictor.
- The metafeatures for each model from IntLIM.
- The phenotype / outcome column name in the input data.
- The type of the outcome / phenotype ("numeric" or "categorical").
- The analyte type used as the outcome variable for each model (1 or 2).
- The analyte type used as the independent variable for each model (1 or 2).
}
\section{Slots}{

\describe{
\item{\code{metaFeatures}}{A list of calculated meta-features for each sample}

\item{\code{modelProperties}}{A data frame that includes model information, i.e. R^2,
interaction term p-value, and coefficients.}

\item{\code{inputData}}{An IntLIMData object that includes slots for the sample data,
the analyte data, and the analyte meta data.}
}}

