% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_ComputeMetaFeatures.R
\name{GetPredictorIDs}
\alias{GetPredictorIDs}
\title{Function for obtaining database identifiers for each predictor.}
\usage{
GetPredictorIDs(preds, inputData, colIdInd, colIdOut)
}
\arguments{
\item{inputData}{Input data, which may include database ID mappings for
the analytes.}

\item{colIdInd}{The ID of the column that has the analyte ID's for the
independent variable. If blank, then the existing ID's are used.}

\item{colIdOut}{The ID of the column that has the analyte ID's for the
outcome variable. If blank, then the existing ID's are used.}

\item{predictions}{Prediction data frame, where rows are samples, and
columns are predictors.}
}
\description{
Function for obtaining database identifiers for each predictor.
}
