% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{InitializeGraphLearningModel}
\alias{InitializeGraphLearningModel}
\title{Find edges that share nodes and add them to a data frame.}
\usage{
InitializeGraphLearningModel(
  modelInputs,
  importance,
  iterations = 1000,
  convergenceCutoff = 0.001,
  modelType = "importance",
  stype.class = "numeric",
  learningRate = 0.2,
  activationType = "none",
  optimizationType = "SGD"
)
}
\arguments{
\item{modelInputs}{An object of the ModelInput class.}

\item{importance}{Calculated importance metrics.}

\item{iterations}{Maximum number of iterations. Default is 1,000.}

\item{convergenceCutoff}{Cutoff for convergence. Default is 0.001.}

\item{modelType}{The type of model the user wishes to learn. Users may
select from the following:
- "importance": A linear combination of predictors, where each predictor's
  weight is learned as a function of importance metrics.}

\item{stype.class}{One of either "character" or "numeric". Default is "numeric".}

\item{learningRate}{Learning rate to use during training. Default is 0.2}

\item{activationType}{Activation function. May be "softmax", "sigmoid", 
"tanh", or "none". Default is "none", meaning stype.class is numeric.}

\item{optimizationType}{Type of optimization. May be "SGD", "momentum",
"adagrad", or "adam". Default is "SGD".}
}
\description{
Find edges that share nodes and add them to a data frame.
}
