% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ObtainCompositeModels}
\alias{ObtainCompositeModels}
\title{Obtain the list of pairs in a composite model.}
\usage{
ObtainCompositeModels(pairsInEachPredictor, importantModels)
}
\arguments{
\item{pairsInEachPredictor}{A list of pairs contained within each predictor.}

\item{importantModels}{A list of all pairs that were found to be important in the
previous layer.}
}
\description{
Obtain the list of pairs in a composite model.
}
