% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{ObtainSubgraphNeighborhoods}
\alias{ObtainSubgraphNeighborhoods}
\title{Obtain the list of pairs in a composite model.}
\usage{
ObtainSubgraphNeighborhoods(modelInput, percentOverlapCutoff = 100)
}
\arguments{
\item{modelInput}{A ModelInput object}

\item{percentOverlapCutoff}{Neighborhoods that have this percent overlap
with another neighborhood will be removed from consideration. Default is 100.}
}
\value{
A list of sets, where each set is a neighborhood of nodes.
}
\description{
Obtain the list of pairs in a composite model.
}
