% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_LearnGraphPredictionModel.R
\name{OptimizeMetaFeatureCombo}
\alias{OptimizeMetaFeatureCombo}
\title{Optimize the combination of predictors by metafeatures alone (in other words,
exclude pooling and combine in a single layer using a linear combination).}
\usage{
OptimizeMetaFeatureCombo(
  modelResults,
  verbose = TRUE,
  modelRetention = "stringent",
  useCutoff = FALSE
)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{verbose}{Whether to print results as you run the model.}

\item{modelRetention}{Strategy for model retention. "stringent" (the default)
retains only models that improve the prediction score. "lenient" also retains models that
neither improve nor reduce the prediction score.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}
}
\value{
A ModelResults object, with all of the tracking information from each
iteration filled in.
}
\description{
Optimize the combination of predictors by metafeatures alone (in other words,
exclude pooling and combine in a single layer using a linear combination).
}
