% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_LearnGraphPredictionModel.R
\name{OptimizeMetaFeatureCombo}
\alias{OptimizeMetaFeatureCombo}
\title{Optimize the combination of predictors by metafeatures alone (in other words,
exclude pooling and combine in a single layer using a linear combination).}
\usage{
OptimizeMetaFeatureCombo(
  modelResults,
  verbose = TRUE,
  pruningMethod = "odds.ratio",
  binCount = 10,
  margin = 0.1,
  includeVarianceTest = FALSE
)
}
\arguments{
\item{modelResults}{An object of the ModelResults class.}

\item{verbose}{Whether to print results as you run the model.}

\item{pruningMethod}{Set to "information.gain", "odds.ratio", or "error.t.test"}

\item{binCount}{The number of bins to divide your original data into. Default is 10.
Used in information gain pruning only.}

\item{margin}{The margin of error for a prediction to be considered "correct".
Default is 0.1. Used in odds ratio pruning only.}

\item{includeVarianceTest}{Scale the t-score by the f-statistic (the ratio of variances).
Only applicable when the pruning method is error.t.test. Default is FALSE.}
}
\description{
Optimize the combination of predictors by metafeatures alone (in other words,
exclude pooling and combine in a single layer using a linear combination).
}
