% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotLineGraph}
\alias{PlotLineGraph}
\title{Plot the line graph for each sample with nodes colored according to prediction.
Include a color scale for the predictions.}
\usage{
PlotLineGraph(
  modelResults,
  subset = c(),
  saveInDir = "",
  truncateTo = 2,
  weights = FALSE,
  analytes = c(),
  includeLabels = TRUE,
  cutoffs = c(0, 0),
  vertexSize = 10,
  sampSubset = c()
)
}
\arguments{
\item{modelResults}{A ModelResults object.}

\item{subset}{The subset of predictors to include. If c(), then all nodes
will be plotted.}

\item{saveInDir}{Directory where file should be saved. If "", then the output
is plotted without being saved. Default is ""}

\item{truncateTo}{Analyte names are truncated to the first "truncateTo" characters.
Default is 4. If -1, names are not truncated.}

\item{weights}{Whether or not to encode weights in the opacity of the nodes.
Defaults to FALSE.}

\item{analytes}{List of analytes to plot. This is used if one
wishes to focus on a subset of analytes If c(), then all vertices are plotted.
Default is c().}

\item{includeLabels}{whether or not to include labels. Defaults to TRUE.}

\item{cutoffs}{Cutoff weight values, which can be included for visibility.
Default is c(0,0), or no cutoff.}

\item{vertexSize}{Vertex size to use in display.}

\item{sampSubset}{Samples to plot. If c(), all samples are plotted.}
}
\description{
Plot the line graph for each sample with nodes colored according to prediction.
Include a color scale for the predictions.
}
