% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotSubspaceClusteringDendrogram}
\alias{PlotSubspaceClusteringDendrogram}
\title{Plots a dendrogram of the optimal subspace clustering.}
\usage{
PlotSubspaceClusteringDendrogram(optimalClustering)
}
\arguments{
\item{optimalClustering}{The output of FindOptimalSubspaceClustering.}
}
\description{
Plots a dendrogram of the optimal subspace clustering.
}
