% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingfunctions.R
\name{PlotSubspaceClusteringHeatmap}
\alias{PlotSubspaceClusteringHeatmap}
\title{Plots a heatmap of the optimal subspace clustering.}
\usage{
PlotSubspaceClusteringHeatmap(
  inputData,
  eigStep = 10,
  alphaMin = 0,
  alphaMax = 1,
  alphaStep = 0.1
)
}
\arguments{
\item{optimalClustering}{The output of FindOptimalSubspaceClustering.}
}
\description{
Plots a heatmap of the optimal subspace clustering.
}
