% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_PredictUsingLearnedModel.R
\name{Predict}
\alias{Predict}
\title{Run a prediction on new data using the graph learning model.}
\usage{
Predict(
  pairs,
  inputData,
  metafeatures,
  model,
  minCutoff,
  maxCutoff,
  useCutoff = FALSE,
  useActivation = TRUE
)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{inputData}{The input testing data, which should be of an 
IntLimData class type.}

\item{metafeatures}{The metafeatures calculated using GetTestMetaFeatures()}

\item{model}{The learned model, an object of the modelResults class.}

\item{minCutoff}{Mininum cutoff for the prediction.}

\item{maxCutoff}{Maximum cutoff for the prediction.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}

\item{useActivation}{Whether or not to use the activation function if the phenotype
to predict is a factor. Default is TRUE. We set to FALSE during training.}
}
\description{
Run a prediction on new data using the graph learning model.
}
