% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_PredictUsingLearnedModel.R
\name{PredictTesting}
\alias{PredictTesting}
\title{Run a prediction on new data using the graph learning model.}
\usage{
PredictTesting(inputData, metafeatures, model)
}
\arguments{
\item{inputData}{The input testing data, which should be of an 
IntLimData class type.}

\item{metafeatures}{The metafeatures calculated using GetTestMetaFeatures()}

\item{model}{The learned model, an object of the modelResults class.}
}
\description{
Run a prediction on new data using the graph learning model.
}
