% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{PrunePredictors}
\alias{PrunePredictors}
\title{Given multiple composite predictors, prune the predictors that are not needed.}
\usage{
PrunePredictors(
  compositeSubgraphs,
  previousModels,
  modelResults,
  verbose = FALSE,
  makePlots = FALSE,
  pruningMethod = "information.gain"
)
}
\arguments{
\item{compositeSubgraphs}{A list of pairs to include in the composite model.}

\item{previousModels}{A list of the previous models that were consolidated.}

\item{modelResults}{A ModelResults object.}

\item{verbose}{Whether or not to print out each step.}

\item{makePlots}{Whether or not to plot the pruned model at each step.}

\item{pruningMethod}{Set to "information.gain", "odds.ratio", or "error.t.test"}
}
\value{
A list of informative pairs
}
\description{
Given multiple composite predictors, prune the predictors that are not needed.
}
