% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{PrunePredictors}
\alias{PrunePredictors}
\title{Given all predictors, find only the important predictors.}
\usage{
PrunePredictors(
  compositeSubgraphs,
  IntLIMresults,
  modelResults,
  inputData,
  verbose = FALSE
)
}
\arguments{
\item{IntLIMresults}{The filtered results obtained from the IntLIM function
ProcessRe}

\item{modelResults}{A ModelResults object.}

\item{inputData}{An object of type IntLimData.}

\item{verbose}{Whether or not to print out each step.}

\item{pairs}{A list of pairs to include in the composite model.}
}
\value{
A list of informative pairs
}
\description{
Given all predictors, find only the important predictors.
}
