% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{PrunePredictors}
\alias{PrunePredictors}
\title{Given multiple composite predictors, prune the predictors that are not needed.}
\usage{
PrunePredictors(
  compositeSubgraphs,
  previousModels,
  modelResults,
  verbose = FALSE,
  makePlots = FALSE,
  pruningMethod = "odds.ratio",
  binCount = 10,
  margin = 0.1,
  includeVarianceTest = includeVarianceTest,
  tolerance = 1e-05,
  modelRetention = "stringent",
  minCutoff,
  maxCutoff,
  useCutoff = FALSE
)
}
\arguments{
\item{compositeSubgraphs}{A list of pairs to include in the composite model.}

\item{previousModels}{A list of the previous models that were consolidated.}

\item{modelResults}{A ModelResults object.}

\item{verbose}{Whether or not to print out each step.}

\item{makePlots}{Whether or not to plot the pruned model at each step.}

\item{pruningMethod}{Set to "information.gain", "odds.ratio", or "error.t.test"}

\item{binCount}{The number of bins to divide your original data into. Default is 10.
Used in information gain pruning only.}

\item{margin}{The margin of error for a prediction to be considered "correct".
Default is 0.1. Used in odds ratio pruning only.}

\item{includeVarianceTest}{Scale the t-score by the f-statistic (the ratio of variances).
Only applicable when the pruning method is error.t.test. Default is FALSE.}

\item{tolerance}{Tolerance factor when computing equality of two numeric values.}

\item{modelRetention}{Strategy for model retention. "stringent" (the default)
retains only models that improve the prediction score. "lenient" also retains models that
neither improve nor reduce the prediction score.}

\item{minCutoff}{Mininum cutoff for the prediction.}

\item{maxCutoff}{Maximum cutoff for the prediction.}

\item{useCutoff}{Whether or not to use the cutoff for prediction. Default is FALSE.}
}
\value{
A list of informative pairs
}
\description{
Given multiple composite predictors, prune the predictors that are not needed.
}
