% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositemodelfunctions.R
\name{PrunePredictors}
\alias{PrunePredictors}
\title{Given all predictors, find only the important predictors.}
\usage{
PrunePredictors(pairs, IntLIMresults, modelResults, inputData)
}
\arguments{
\item{pairs}{A list of pairs to include in the composite model.}

\item{IntLIMresults}{The filtered results obtained from the IntLIM function
ProcessRe}

\item{modelResults}{A ModelResults object.}

\item{inputData}{An object of type IntLimData.}
}
\value{
A list of informative pairs
}
\description{
Given all predictors, find only the important predictors.
}
