% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwisepredictionfunctions.R
\name{RunPairwisePrediction}
\alias{RunPairwisePrediction}
\title{Given each significant pairwise model and the input data, predict the phenotype
for each sample. Recall that IntLIM models take the following form, where a_i
and a_j are a pair of analytes.
a_i ~ beta0 + beta1(a_j) + beta2(phenotype) + beta3(a_j:phenotype) + beta4...n(covariates)
Therefore, to predict phenotype given the betas learned by IntLIM, we use the
following model:
p ~ (a_i - (beta0 + beta1(a_j) + beta4...n(covariates)) / (beta2 + beta3(a_j))}
\usage{
RunPairwisePrediction(
  inputResults,
  inputData,
  stype = "",
  covar = c(),
  independentVarType = 2,
  outcomeType = 1
)
}
\arguments{
\item{inputResults}{The data frame of filtered results from IntLIM
model and processing results (output of ProcessResults()). All results must
include learned covariate weights (i.e. must be run with save.covar.pvals = TRUE)}

\item{inputData}{An IntLimData object that includes the input.}

\item{stype}{The phenotype (outcome) to predict. This can be either a categorical
or numeric outcome.}

\item{covar}{The clinical covariates to include in the model. These should be the same
covariates that were included when running the IntLIM linear models.}

\item{independentVarType}{The independent variable type (1 or 2)}

\item{outcomeType}{The outcome type (1 or 2)}
}
\value{
A data frame of predictions, where each column is a sample and each
row is a predictor.
}
\description{
Given each significant pairwise model and the input data, predict the phenotype
for each sample. Recall that IntLIM models take the following form, where a_i
and a_j are a pair of analytes.
a_i ~ beta0 + beta1(a_j) + beta2(phenotype) + beta3(a_j:phenotype) + beta4...n(covariates)
Therefore, to predict phenotype given the betas learned by IntLIM, we use the
following model:
p ~ (a_i - (beta0 + beta1(a_j) + beta4...n(covariates)) / (beta2 + beta3(a_j))
}
