% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_PredictUsingLearnedModel.R
\name{SplitTrainingAndTesting}
\alias{SplitTrainingAndTesting}
\title{Splits input data into training and testing sets given a specified
set of testing samples. Alternatively, separate training and testing
sets can be read in prior to training. However, this function is a
useful helper for cross-validation.}
\usage{
SplitTrainingAndTesting(inputData, testingSamples)
}
\arguments{
\item{inputData}{An object of the IntLimData class.}

\item{testingSamples}{Calculated importance metrics.}
}
\value{
A list of two IntLimData objects: a training set and a testing set.
}
\description{
Splits input data into training and testing sets given a specified
set of testing samples. Alternatively, separate training and testing
sets can be read in prior to training. However, this function is a
useful helper for cross-validation.
}
